Wrote support for multiple control schemes.
I know a lot of people preferred last year's control scheme, so I decided to code it back with a match statement.

One match statement is the arcade control set that Esteban wants to use.

Another match statement is the control set that is mostly standardized across FRC for Arcade.

One control set is last year's.

The last one is tank drive.